IS-1619: Changed name of button
